+++ CHANGELOG.md
@@ -8,6 +8,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
 - utf-8 as default encoding
 - `CreateContactBag::withPhone` marked as deprecated
 
+### Removed
+- `SendSmsBag::setIdx`
+
 ## [2.6.1] - 2020-05-14
 ### Fixed
 - undelivered messages contact field support